YEAH

+++ hello.cu
@@ -24,9 +24,9 @@ int main()
     //init
     printf("Init...\n");
     for(int i = 0; i < n; i++){
-        x[i] = (float) (i * 2.5) % 13;
-        y[i] = (float) (i * 1.9) % 25;
-        z[i] = (float) 0;
+        x[i] = (i * 2.5) % 13;
+        y[i] = (i * 1.9) % 25;
+        z[i] = 0;
     }
     printf("Init complete\n");
